cifsadmin: perform sanity checks before user_del
Check if the user we are about to delete is:
- still in global guest user
- is still required to exist by any of net shares:
  - admin users list
  - write users list
  - valid users list

Abort user deletion in any of these cases.

Otherwise, we can drop user X, which is given admin permission
on share Y, then someday add a new user X, which will get admin
permission on share Y automatically, even though we might not
want to actually give that new user admin permissions.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>