cifsd-tools: use a constant-expression for hex_dump buffer size
ISO C90 requires an array size to be not a constant value,
but a constant expression.

So replace that nice 'const int WIDTH = 160' with a macro; to shut
up^W^W make gcc happy.

Discovered by -Wvla, which is garbage: misleading name, misleading
warning. It has nothing to do with "variable length" and whether
the compiler can figure it out at build time.

Signed-off-by: Rosen Penev <rosenp@gmail.com>
Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>